% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocialInductionModels.R
\name{socialInduction}
\alias{socialInduction}
\title{Social induction model}
\usage{
socialInduction(obsNet = startingNet, Pn, Pr, Pb, iter)
}
\arguments{
\item{obsNet}{Starting network. This must be a non-weighted and non-directed igraph network.}

\item{Pn}{Probability of forming ties with a mother's partner.}

\item{Pr}{Probability of forming ties with a non mother's partner.}

\item{Pb}{Probability of forming ties with mother.}

\item{iter}{Number of iteration, where each iteration removes one node and introduces a new one.}
}
\description{
This function runs the social induction model of network formation (see Ilany et al. 2016)
}
\examples{
startingNet = erdos.renyi.game(n=25, type = c("gnm"), p.or.m = 45)
predictedNet<-socialInduction(startingNet, Pn=0.4, Pr=0.1, Pb=1, iter=100)
plot(startingNet)
plot(predictedNet)

}
\references{
Ilany A, Akcay E. Social inheritance can explain the structure of animal social networks. Nature Communications. 2016;7:12084. doi:10.1038/ncomms12084.
}
\keyword{Formation,}
\keyword{Network}
\keyword{models}
\keyword{social}
