% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocialInductionModels.R
\name{socialInductionWeighted}
\alias{socialInductionWeighted}
\title{Social induction model with weights}
\usage{
socialInductionWeighted(obsNet = startingNet, Pn, Pr, Pb, En1, En2, Er1, Er2,
  maxE, iter)
}
\arguments{
\item{obsNet}{Starting network. This must be a weighted and non-directed igraph network.}

\item{Pn}{Probability of forming ties with a mother's partner.}

\item{Pr}{Probability of forming ties with a non mother's partner.}

\item{Pb}{Probability of forming ties with mother.}

\item{En1}{Alpha parameter for effort into a mother's partner. Where effort is defined by a beta distribution (0,1) using alpah and beta.}

\item{En2}{Beta parameter for effort into a mother's partner. Where effort is defined by a beta distribution (0,1) using alpah and beta.}

\item{Er1}{Alpha parameter for effort into a non mother's partner. Where effort is defined by a beta distribution (0,1) using alpah and beta.}

\item{Er2}{Beta parameter for effort into a non mother's partner. Where effort is defined by a beta distribution (0,1) using alpah and beta.}

\item{maxE}{Maximum grooming effort used to scale the beta distribution. This choice should represent an upper limit of effort.}

\item{iter}{Number of iteration, where each iteration removes one node and introduces a new one.}
}
\description{
This function runs the social induction model of network formation (see Ilany et al. 2016), with the addition of weighted edges.
}
\examples{
startingNet = erdos.renyi.game(n=10, type = c("gnm"), p.or.m = 15)
E(startingNet)$weight<-c(2,25,1,1,1,1,4,5,6,9,2,1,2,1,2)
predWeightedNet<-socialInductionWeighted(startingNet, Pn=0.4,Pr=0.1,Pb=1,En1=0.1,En2=4,Er1=0.1,Er2=4,maxE=100,iter=100)
plot(startingNet, edge.width=E(startingNet)$weight)
plot(predWeightedNet, edge.width=E(predWeightedNet)$weight)


}
\keyword{Formation,}
\keyword{Network}
\keyword{models}
\keyword{social}
